prep for yet another refactoring, macOS temporarily disabled; updates #32'

+++ prototype2/gdgem/dg_impl/NMXCluster.h
@@ -5,8 +5,8 @@
 #include <gdgem/nmx/Eventlet.h>
 
 struct ClusterNMX {
-  int size;
-  int adc;
+  size_t size;
+  uint64_t adc;
   double position;
   double time;
   bool clusterXAndY;
